Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
elasticsearch==8.12.0
aiohttp
passlib
bcrypt==4.0.1
python-jose
python-multipart
email-validator==2.2.0
pydantic
psycopg2-binary
python-dotenv
requests
json
pathlib
feedparser
asyncio